refactor(viz): merge SHistogram, SHistogramCursor and SHistogramValue together
Since we removed SCurveHistogram, there was no good reason to keep those three adaptors separately.

Doing that, we improved the computation speed of the image histogram by a factor of 10. Computing the histogram is now so fast that it is useless to have a dedicated data for it. We removed all the ecosystem around it and the only service that uses it so far, the SHistogram adaptor, computes the histogram dynamically. On top of that, this allows to change the bins width dynamically which is very handy.

We also added an axis for the image histogram that can scale with the graph. This required a refactor of the usage of the viewport in the 2D scene. Now the scene no longer holds a viewport, but can only update its viewing rect through a dedicated function available to the adaptors. The adaptors that do use the viewport now all explicitly declare it, with the correct access type (in or inout).  Last, we scale the transfer function viewport range automatically from image min and max intensities.

Many warnings were fixed along, notably because of the coexistence of float and double variables. The floating point precision was set to double everywhere to comply with Qt API.

+++ libs/viz/scene2d/data/Axis.hpp
@@ -1,6 +1,6 @@
 /************************************************************************
  *
- * Copyright (C) 2009-2021 IRCAD France
+ * Copyright (C) 2009-2022 IRCAD France
  * Copyright (C) 2012-2021 IHU Strasbourg
  *
  * This file is part of Sight.
@@ -56,16 +56,16 @@ public:
     }
 
     /// Get m_origin attribute.
-    float getOrigin() const;
+    double getOrigin() const;
 
     /// Set m_origin attribute.
-    void setOrigin(float origin);
+    void setOrigin(double origin);
 
     /// Get m_scale attribute.
-    float getScale() const;
+    double getScale() const;
 
     /// Set m_scale attribute.
-    void setScale(float scale);
+    void setScale(double scale);
 
     /// Get m_scaleType attribute.
     ScaleType getScaleType() const;
@@ -76,7 +76,7 @@ public:
 private:
 
     /// The origin and scale.
-    float m_origin, m_scale;
+    double m_origin, m_scale;
 
     /// The scaleType (LINEAR/LOG).
     ScaleType m_scaleType;
@@ -84,28 +84,28 @@ private:
 
 //-----------------------------------------------------------------------------
 
-inline float Axis::getOrigin() const
+inline double Axis::getOrigin() const
 {
     return m_origin;
 }
 
 //-----------------------------------------------------------------------------
 
-inline void Axis::setOrigin(float origin)
+inline void Axis::setOrigin(double origin)
 {
     m_origin = origin;
 }
 
 //-----------------------------------------------------------------------------
 
-inline float Axis::getScale() const
+inline double Axis::getScale() const
 {
     return m_scale;
 }
 
 //-----------------------------------------------------------------------------
 
-inline void Axis::setScale(float scale)
+inline void Axis::setScale(double scale)
 {
     m_scale = scale;
 }
